dom/virtualfilesystem/test/mochitest/test_virtualfilesystem_mount.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1026350
-->
<head>
<meta charset="utf-8">
<title>Test Inputport Connection Event</title>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1187723">Test FileSystemProvider Event
</a>
<script type="application/javascript;version=1.8">

'use strict';

SimpleTest.waitForExplicitFinish();

function testMount() {
  return new Promise(function(resolve, reject) {
    var fileSystemProvider = navigator.fileSystemProvider;
    var _fileSystemId = "dummyId";
    var _displayName = "dummyFileSystem";
    var _writable = true;
    var _openedFilesLimit = 10;
    fileSystemProvider.mount({fileSystemId: _fileSystemId,
                              displayName: _displayName,
                              writable: _writable,
                              openedFilesLimit: _openedFilesLimit}).then(
      function() {
        ok(true, "mount ok.");
        var fileSystemInfo = fileSystemProvider.get(_fileSystemId);
        ok(fileSystemInfo, "got file system info");
        is(fileSystemInfo.fileSystemId, _fileSystemId, "The fileSystemId name should be the same.");
        is(fileSystemInfo.displayName, _displayName, "The displayName should be the same.");
        is(fileSystemInfo.writable, _writable, "The writable should be the same.");
        is(fileSystemInfo.openedFilesLimit, _openedFilesLimit, "The openedFilesLimit should be the same.");

        fileSystemProvider.unmount({fileSystemId:_fileSystemId}).then(
          function() {
            ok(true, "unmount ok.");
            try {
              var info = fileSystemProvider.get(_fileSystemId);
            } catch (e) {
              ok(true, "can't get file system now.");
              resolve();
            }
          },
          function(aError) {
            ok(false, "Fail to unmount: " + aError);
            resolve();
          }
        );
      },
      function(aError) {
        ok(false, "Fail to mount: " + aError);
        resolve();
      }
    );
  });
}


function runTest() {
  ok(navigator.fileSystemProvider, 'should have navigator.fileSystemProvider');

  testMount()
  .then(function() {
    info('test finished');
    SimpleTest.finish();
  });
}

SpecialPowers.pushPrefEnv({"set": [["device.storage.enabled", true],
                                    //to ignore app scope check.
                                   ["dom.ignore_webidl_scope_checks", true]]}, function() {
  SpecialPowers.pushPermissions(
    [{'type': 'filesystemprovider', 'allow': true, 'context': document}], runTest);
});

</script>
</pre>
</body>
</html>
